[FIX] 파일정리 및 코드정리(#81)
