--- Plu-iOS/Plu-iOS/Scenes/MyPage/Cell/MyPageGeneralTableViewCell.swift
@@ -1,66 +1,0 @@
-//
-//  MyPageNavigationTableViewCell.swift
-//  Plu-iOS
-//
-//  Created by uiskim on 2023/12/05.
-//  Copyright (c) 2023 MyPageNavigation. All rights reserved.
-//
-
-import UIKit
-
-import SnapKit
-
-final class MyPageGeneralTableViewCell: UITableViewCell {
-    
-    private let cellTitle: UILabel = {
-        let label = UILabel()
-        label.font = .suite(.body1M)
-        label.textColor = .designSystem(.black)
-        return label
-    }()
-    
-    private let rightArrow: UIImageView = {
-        let imageView = UIImageView(image: ImageLiterals.MyPage.arrowRightSmall900)
-        imageView.contentMode = .scaleAspectFit
-        return imageView
-    }()
-
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        setUI()
-        setHierarchy()
-        setLayout()
-    }
-    
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    func configureUI(_ input: String) {
-        self.cellTitle.text = input
-    }
-}
-
-private extension MyPageGeneralTableViewCell {
-    func setUI() {
-        self.contentView.backgroundColor = .designSystem(.white)
-    }
-    
-    func setHierarchy() {
-        self.contentView.addSubviews(cellTitle, rightArrow)
-    }
-    
-    func setLayout() {
-        cellTitle.snp.makeConstraints { make in
-            make.centerY.equalToSuperview()
-            make.leading.equalToSuperview().inset(20)
-        }
-        
-        rightArrow.snp.makeConstraints { make in
-            make.centerY.equalToSuperview()
-            make.trailing.equalToSuperview().inset(14)
-            make.size.equalTo(24)
-        }
-    }
-}
