+++ Plu-iOS/Plu-iOS/Scenes/MyPage/ViewController/MyPageViewController.swift
@@ -14,6 +14,11 @@ import SnapKit
 
 final class MyPageViewController: UIViewController {
     
+    private enum MypageType { case profile, alarm, info, version, user }
+    
+    // 임시변수 추후 뷰모델로 이동
+    var isOn = false
+    
     let headerTapped = PassthroughSubject<MypageNavigationType, Never>()
     let faqCellTapped = PassthroughSubject<MypageNavigationType, Never>()
     let backButtonTapped = PassthroughSubject<MypageNavigationType, Never>()
@@ -22,7 +27,6 @@ final class MyPageViewController: UIViewController {
     let alarmSwitchTapped = PassthroughSubject<MypageNavigationType, Never>()
     let openSourceCellTapped = PassthroughSubject<MypageNavigationType, Never>()
     let privacyCellTapped = PassthroughSubject<MypageNavigationType, Never>()
-
     let viewWillAppearSubject = PassthroughSubject<Void, Never>()
     let switchOnSubject = PassthroughSubject<Void, Never>()
     var cancelBag = Set<AnyCancellable>()
@@ -34,7 +38,7 @@ final class MyPageViewController: UIViewController {
         .setLeftButton(type: .back)
     
     private let myPageTableView = UITableView(frame: .zero, style: .grouped)
-    let renderer = Renderer(adapter: UITableViewAdapter(), updater: UITableViewUpdater())
+    private let renderer = Renderer(adapter: UITableViewAdapter(), updater: UITableViewUpdater())
     
     init(viewModel: some MyPageViewModel & MyPagePresentable) {
         self.viewModel = viewModel
@@ -56,43 +60,35 @@ final class MyPageViewController: UIViewController {
         render()
     }
     
-    func setTableView() {
-        myPageTableView.separatorStyle = .none
-        renderer.target = myPageTableView
-    }
-    
     override func viewWillDisappear(_ animated: Bool) {
         self.tabBarController?.tabBar.isHidden = false
     }
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
+        self.tabBarController?.tabBar.isHidden = true
         self.viewWillAppearSubject.send(())
     }
     
-    enum MypageType {
-        case profile, alarm, info, version, user
-    }
+
     func render() {
         renderer.render {
-            
             Section(id: MypageType.profile) {
                 ProfileItem(nickname: "Plu님") {
-                    print("프로필눌림")
+                    self.headerTapped.send(.header)
                 }
             }
             
             Section(id: MypageType.alarm) {
                 RoundHeadItem()
-                
-                AlarmSettingItem(isOn: true) { type in
+                AlarmSettingItem(isOn: self.isOn) { type in
                     switch type {
                     case .alarmAccept:
                         self.alarmSwitchTapped.send(.alarm)
                     case .alarmReject:
                         self.goToSettingPage { _ in
-                            print("✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅✅")
-                            
+                            self.isOn = false
+                            self.render()
                         }
                     }
                 }
@@ -101,33 +97,28 @@ final class MyPageViewController: UIViewController {
             Section(id: MypageType.info) {
                 SpaceItem(16)
                 MyPageItem(title: "FAQ") {
-                    print("FAQ가 눌림")
+                    self.faqCellTapped.send(.faq)
                 }
-                
                 MyPageItem(title: "오픈소스 라이브러리") {
-                    print("오픈소스 라이브러리가 눌림")
+                    self.openSourceCellTapped.send(.openSource)
                 }
-                
                 MyPageItem(title: "개인정보 보호 및 약관") {
-                    print("개인정보 보호 및 약관이 눌림")
+                    self.privacyCellTapped.send(.privacy)
                 }
             }
             
             Section(id: MypageType.version) {
                 SpaceItem(16)
-                
                 AppversionItem(version: "1.0.0")
             }
             
             Section(id: MypageType.user) {
                 SpaceItem(16)
-                
                 MyPageItem(title: "로그아웃") {
-                    print("로그아웃이 눌림")
+                    self.logoutCellTapped.send(())
                 }
-                
                 MyPageItem(title: "탈퇴하기") {
-                    print("탈퇴하기가 눌림")
+                    self.resignCellTapped.send(.resign)
                 }
             }
         }
@@ -191,109 +182,9 @@ private extension MyPageViewController {
             }
             .store(in: &cancelBag)
     }
-
-}
-
-extension MyPageViewController: UITableViewDataSource {
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return self.viewModel.tableData.count
-    }
     
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return self.viewModel.tableData[section].count
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        switch self.viewModel.tableData[indexPath.section][indexPath.row] {
-        case .alarm(let data):
-            let cell = MyPageAlarmTableViewCell.dequeueReusableCell(to: tableView)
-            
-            cell.alarmSwitchTypeSubject
-                .sink { [weak self] type in
-                    switch type {
-                    case .alarmAccept:
-                        self?.alarmSwitchTapped.send(.alarm)
-                    case .alarmReject:
-                        self?.goToSettingPage { _ in 
-                            cell.alarmSwitch.setOn(false, animated: false)
-                        }
-                    }
-                }
-                .store(in: &cell.cancelBag)
-            
-            self.switchOnSubject.sink {
-                cell.alarmSwitch.isOn = true
-            }
-            .store(in: &cell.cancelBag)
-            
-            cell.configureUI(data)
-            return cell
-        case .info(let type):
-            let cell = MyPageGeneralTableViewCell.dequeueReusableCell(to: tableView)
-            cell.configureUI(type.title)
-            return cell
-        case .appVersion(let data):
-            let cell = MyPageAppVersionTableViewCell.dequeueReusableCell(to: tableView)
-            cell.configureUI(data)
-            return cell
-        case .exit(let type):
-            let cell = MyPageGeneralTableViewCell.dequeueReusableCell(to: tableView)
-            cell.configureUI(type.title)
-            return cell
-        }
-    }
-    
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        switch self.viewModel.tableData[indexPath.section][indexPath.row] {
-        case .alarm: return MyPageCellHeight.alarmCellHeight
-        case .info, .exit: return MyPageCellHeight.infoCellHeight
-        case .appVersion: return MyPageCellHeight.appVersionCellHeight
-        }
-    }
-}
-
-extension MyPageViewController: UITableViewDelegate, MyPageHeaderDelgate {
-    func headerDidTapped() {
-        self.headerTapped.send(.header)
-    }
-    
-    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        guard let section = self.viewModel.tableData[section].first else { return UIView() }
-        if case .alarm = section {
-            return MyPageRoundSectionHeaderView.dequeueReusableSectionHeaderView(to: tableView)
-        }
-        let seperateView = UIView()
-        seperateView.backgroundColor = .designSystem(.background)
-        return seperateView
-    }
-    
-    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        guard let section = self.viewModel.tableData[section].first else { return 0 }
-        if case .alarm = section {
-            return MyPageSectionHeaderHeight.alarmSectionHeaderHeight
-        }
-        return MyPageSectionHeaderHeight.seperateSectionHeaderHeight
-    }
-    
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let cellType = self.viewModel.tableData[indexPath.section][indexPath.row]
-        if case .info(let type) = cellType {
-            switch type {
-            case .faq:
-                self.faqCellTapped.send(type.changeToMyPageNavigation)
-            case .openSource:
-                self.openSourceCellTapped.send(type.changeToMyPageNavigation)
-            case .privacy:
-                self.openSourceCellTapped.send(type.changeToMyPageNavigation)
-            }
-        }
-        if case .exit(let type) = cellType {
-            switch type {
-            case .logout:
-                self.logoutCellTapped.send(())
-            case .resign:
-                self.resignCellTapped.send(type.changeToMyPageNavigation)
-            }
-        }
+    func setTableView() {
+        myPageTableView.separatorStyle = .none
+        renderer.target = myPageTableView
     }
 }
